Add disk resizing and ip address reporting
When the disk image is copied to the new image, it is resized
to the size of DISK_GB.

The vm is booted with appropriate VCPU.

The assigned ip address is found with virsh domifadd <uuid> and
reported back to etcd.

All of this stuff is done in a very clumsy and not at all
error proof fashion. But it is a nice educational exercise.