Add simple config file for compute
This makes it easier for there to be compute nodes which are
remote from where the placement and etcd servers are running.

The etcd key takes any parameters which may be passed to the
`etcd3.Client` constructor. The placement key takes a URL.

To make talking to placement cleander, a PrefixedSession
subclass is created which prepends the placement url to
placement requests.