Add a command to delete an instance
If the args 'destroy' and an instance uuid are passed to
schedule.py, the current allocations for that instance are
retrieved, updated to have no allocations, and put to the
etcd key representing the host where the instance is allocated.

The compute.py will reacted to this (in a separate commit).