Destroy instance when incoming allocations empty
If the incoming allocations on the etcd key being watched
by compute.py are empty, destroy and undefine the associated
instance and clear the allocations for the instance in
placement.

Experimenting with this helps to show just how racy the etcd
watching is. If a VM is in the process of being booted or
destroyed, other requests to boot or destroy on the same host
are ignored, without error.

Doing something like forking when the key watch event fires
will help, but it doesn't address a situation where multiple
different instances want attention at the same time.

I suspect etcd3 library can help on this, but it's not been
investigated yet.