Correct caching of disk images
The cachecontrol library can't cache large images [1], which we
work around here by using our own serializer to allow msgpage
to load big files.

It's quite likely this caching is the wrong way to go, and
missing some important details, but it is a useful way to
speed up the experimentation.

[1] https://github.com/ionrock/cachecontrol/issues/200