Correctly handle concurrency
This was a bear to get right. There is some very fancy
handling going on in gprc (used by etcd3) which means
that if you try to use the etcd3 from a forked client
things go awry, with processes blocking and the like.

To resolve this, instead of forking directly, use
a multiprocessing.Pool with async workers that return
either False or an IP address. Those results are then
used to update etcd. This means that all etcd3 interactions
are done in the main process where they don't seem to
conflict with one another.

The end result is that it is relatively straightforward
to launch multiple VMs in a relative hurry and it
basically works.

Things that still need to be tried/addressed:

* what happens when trying to concurrently create more vms
  than there are pool workers?

* the code is in complete disarray now as I flailed about

* the README needs to be updated to reflect this new
  concurrent state of affairs

* This work has made it very clear that real logging is needed
  to accurately debug and info and all that.