Add support for bridging vms to external network
This allows a VM to get DHCP for the LAN and be reached from the
lan. Combined with judicious use of the metadate server, we're
now able to boot a VM and ssh into with a public key.

Assuming the bridge interface is set up correctly. Doing so
is described in BRIDGE.MD