All naming image in eschedule
If the first args is 'resources' an optional second arg names
the URL of image to be used. If it is not already present in
the cache on the selected compute node, it will be fetched.

This now makes it clear that the use of the metadata server
is no good and what we really want is to add user data to
the disk image.