Add an INSTALL.md
This is a first draft of an INSTALL document which expands on
history gathered when making a hypervisor host works. It is
incomplete but should be an improvement on what's in the
README.

It is not yet linked from the README because it is a) incomplete,
b) needs some review and refinement.